+++ src/test/java/madstodolist/UsuarioTest.java
@@ -26,9 +26,6 @@ public class UsuarioTest {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
-    @Autowired
-    private UsuarioService usuarioService;
-
     @Test
     public void crearUsuario() throws Exception {
 
@@ -103,99 +100,6 @@ public class UsuarioTest {
         assertThat(usuario.getNombre()).isEqualTo("Ana García");
     }
 
-
-    @Test
-    @Transactional(readOnly = true)
-    public void servicioLoginUsuario() {
-        // GIVEN
-        // Datos cargados de datos-test.sql
-
-        // WHEN
-
-        UsuarioService.LoginStatus loginStatusOK = usuarioService.login("ana.garcia@gmail.com", "12345678");
-        UsuarioService.LoginStatus loginStatusErrorPassword = usuarioService.login("ana.garcia@gmail.com", "000");
-        UsuarioService.LoginStatus loginStatusNoUsuario = usuarioService.login("pepito.perez@gmail.com", "12345678");
-
-        // THEN
-
-        assertThat(loginStatusOK).isEqualTo(UsuarioService.LoginStatus.LOGIN_OK);
-        assertThat(loginStatusErrorPassword).isEqualTo(UsuarioService.LoginStatus.ERROR_PASSWORD);
-        assertThat(loginStatusNoUsuario).isEqualTo(UsuarioService.LoginStatus.USER_NOT_FOUND);
-    }
-
-    @Test
-    @Transactional
-    public void servicioRegistroUsuario() {
-        // GIVEN
-
-        Usuario usuario = new Usuario("usuario.prueba@gmail.com");
-        usuario.setPassword("12345678");
-
-        // WHEN
-
-        usuarioService.registrar(usuario);
-
-        // THEN
-
-        Usuario usuarioBaseDatos = usuarioRepository.findByEmail("usuario.prueba@gmail.com").orElse(null);
-        assertThat(usuarioBaseDatos).isNotNull();
-        assertThat(usuarioBaseDatos.getPassword()).isEqualTo(usuario.getPassword());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void servicioRegistroUsuarioExcepcionConNullPassword() {
-        // Pasamos como argumento un usario sin contraseña
-        Usuario usuario =  new Usuario("usuario.prueba@gmail.com");
-        usuarioService.registrar(usuario);
-    }
-
-
-    @Test(expected = UsuarioServiceException.class)
-    public void servicioRegistroUsuarioExcepcionConEmailRepetido() {
-        // GIVEN
-        // Datos cargados de datos-test.sql
-
-        // WHEN
-        // Pasamos como argumento un usario con emaii existente en datos-test.sql
-        Usuario usuario =  new Usuario("ana.garcia@gmail.com");
-        usuario.setPassword("12345678");
-        usuarioService.registrar(usuario);
-
-        // THEN
-        // Se produce una excepción comprobada con el expected del test
-    }
-
-    @Test
-    @Transactional
-    public void servicioRegistroUsuarioDevuelveUsuarioConId() {
-        // GIVEN
-
-        Usuario usuario = new Usuario("usuario.prueba@gmail.com");
-        usuario.setPassword("12345678");
-
-        // WHEN
-
-        usuario = usuarioService.registrar(usuario);
-
-        // THEN
-
-        assertThat(usuario.getId()).isNotNull();
-    }
-
-    @Test
-    @Transactional(readOnly = true)
-    public void servicioConsultaUsuarioDevuelveUsuario() {
-        // GIVEN
-        // Datos cargados de datos-test.sql
-
-        // WHEN
-
-        Usuario usuario = usuarioService.findByEmail("ana.garcia@gmail.com");
-
-        // THEN
-
-        assertThat(usuario.getId()).isEqualTo(1L);
-    }
 
     @Test
     public void comprobarIgualdadSinId() {
